Convert `content` to TypeScript.
