+++ packages/agent-kit/src/templating.ts
@@ -7,7 +7,7 @@
 import { Schema } from "@google-labs/breadboard";
 import { Context, LlmContent } from "./context.js";
 
-export { substitute, describeSpecialist };
+export { substitute, describeSpecialist, content };
 
 type SubstituteInputParams = {
   in?: Context[];
@@ -35,6 +35,12 @@ type SpecialistDescriberInputs = {
   $outputSchema: Schema;
   persona?: LlmContent;
   task?: LlmContent;
+};
+
+type ContentInputs = {
+  template: LlmContent;
+  context?: LlmContent[];
+  [param: string]: unknown;
 };
 
 /**
@@ -334,5 +340,217 @@ function describeSpecialist(inputs: unknown) {
     if (!text) return [];
     const matches = text.matchAll(/{{(?<name>[\w-]+)}}/g);
     return Array.from(matches).map((match) => match.groups?.name || "");
+  }
+}
+
+/**
+ * The guts of the "Content" component.
+ */
+function content(starInputs: unknown) {
+  const { template, context, ...inputs } = starInputs as ContentInputs;
+  const params = mergeParams(findParams(template));
+  const values = collectValues(params, inputs);
+
+  return {
+    context: prependContext(context, subContent(template, values)),
+  };
+
+  function prependContext(
+    context: LlmContent[] | undefined,
+    c: LlmContent | string
+  ): LlmContent[] {
+    const content = (isEmptyContent(c) ? [] : [c]) as LlmContent[];
+    if (!context) return [...content] as LlmContent[];
+    if (isLLMContentArray(context)) {
+      // If the last item in the context has a user rule,
+      // merge the new content with it instead of creating a new item.
+      const last = context.at(-1);
+      if (last && last.role === "user") {
+        return [
+          ...context.slice(0, -1),
+          {
+            role: "user",
+            parts: [
+              ...last.parts,
+              ...((content.at(0) as LlmContent)?.parts || []),
+            ],
+          },
+        ];
+      }
+      return [...context, ...content] as LlmContent[];
+    }
+    return content;
+  }
+
+  function isEmptyContent(content: LlmContent | string | undefined) {
+    if (!content) return true;
+    if (typeof content === "string") return true;
+    if (!content.parts?.length) return true;
+    if (content.parts.length > 1) return false;
+    const part = content.parts[0];
+    if (!("text" in part)) return true;
+    if (part.text.trim() === "") return true;
+    return false;
+  }
+
+  function subContent(
+    content: LlmContent | undefined,
+    values: Record<string, unknown>
+  ): LlmContent | string {
+    // If this is an array, optimistically presume this is an LLM Content array.
+    // Take the last item and use it as the content.
+    if (Array.isArray(content)) {
+      content = content.at(-1);
+    }
+    if (!content) return "";
+    return {
+      role: content.role || "user",
+      parts: mergeTextParts(
+        splitToTemplateParts(content).flatMap((part) => {
+          if ("param" in part) {
+            const value = values[part.param];
+            if (typeof value === "string") {
+              return { text: value };
+            } else if (isLLMContent(value)) {
+              return value.parts;
+            } else if (isLLMContentArray(value)) {
+              const last = value.at(-1);
+              return last ? last.parts : [];
+            } else {
+              return { text: JSON.stringify(value) };
+            }
+          } else {
+            return part;
+          }
+        })
+      ),
+    };
+  }
+
+  function findParams(content: LlmContent | undefined): ParamInfo[] {
+    const parts = content?.parts;
+    if (!parts) return [];
+    const results = parts.flatMap((part) => {
+      if (!("text" in part)) return [];
+      const matches = part.text.matchAll(/{{(?<name>[\w-]+)}}/g);
+      return unique(Array.from(matches))
+        .map((match) => {
+          const name = match.groups?.name || "";
+          if (!name) return null;
+          return { name, locations: [{ part, parts }] };
+        })
+        .filter(Boolean);
+    }) as unknown as ParamInfo[];
+    return results;
+  }
+
+  function mergeParams(...paramList: ParamInfo[][]) {
+    return paramList.reduce((acc, params) => {
+      for (const param of params) {
+        const { name, locations } = param;
+        const existing = acc[name];
+        if (existing) {
+          existing.push(...locations);
+        } else {
+          acc[name] = locations;
+        }
+      }
+      return acc;
+    }, {} as ParamLocationMap);
+  }
+
+  function unique<T>(params: T[]): T[] {
+    return Array.from(new Set(params));
+  }
+
+  function mergeTextParts(parts: TemplatePart[]) {
+    const merged = [];
+    for (const part of parts) {
+      if ("text" in part) {
+        const last = merged[merged.length - 1];
+        if (last && "text" in last) {
+          last.text += part.text;
+        } else {
+          merged.push(part);
+        }
+      } else {
+        merged.push(part);
+      }
+    }
+    return merged as LlmContent["parts"];
+  }
+
+  function toId(param: string) {
+    return `p-${param}`;
+  }
+
+  function toTitle(id: string) {
+    const spaced = id?.replace(/[_-]/g, " ");
+    return (
+      (spaced?.at(0)?.toUpperCase() ?? "") +
+      (spaced?.slice(1)?.toLowerCase() ?? "")
+    );
+  }
+
+  /**
+   * Takes an LLM Content and splits it further into parts where
+   * each {{param}} substitution is a separate part.
+   */
+  function splitToTemplateParts(content: LlmContent): TemplatePart[] {
+    const parts = [];
+    for (const part of content.parts) {
+      if (!("text" in part)) {
+        parts.push(part);
+        continue;
+      }
+      const matches = part.text.matchAll(/{{(?<name>[\w-]+)}}/g);
+      let start = 0;
+      for (const match of matches) {
+        const name = match.groups?.name || "";
+        const end = match.index;
+        if (end > start) {
+          parts.push({ text: part.text.slice(start, end) });
+        }
+        parts.push({ param: name });
+        start = end + match[0].length;
+      }
+      if (start < part.text.length) {
+        parts.push({ text: part.text.slice(start) });
+      }
+    }
+    return parts;
+  }
+
+  function collectValues(
+    params: ParamLocationMap,
+    inputs: Record<string, unknown>
+  ) {
+    const values: Record<string, unknown> = {};
+    for (const param in params) {
+      const id = toId(param);
+      const value = inputs[id];
+      if (!value) {
+        const title = toTitle(param);
+        throw new Error(`Missing required parameter: ${title}`);
+      }
+      values[param] = value;
+    }
+    return values;
+  }
+
+  /**
+   * Copied from @google-labs/breadboard
+   */
+  function isLLMContent(nodeValue: unknown): nodeValue is LlmContent {
+    if (typeof nodeValue !== "object" || !nodeValue) return false;
+    if (nodeValue === null || nodeValue === undefined) return false;
+
+    return "parts" in nodeValue && Array.isArray(nodeValue.parts);
+  }
+
+  function isLLMContentArray(nodeValue: unknown): nodeValue is LlmContent[] {
+    if (!Array.isArray(nodeValue)) return false;
+    if (nodeValue.length === 0) return true;
+    return isLLMContent(nodeValue.at(-1));
   }
 }
